Refactor code for s3Upload
Refactor code and moved the aws s3.upload function to lib/ directory
so that we can use it from our API in the future. Required this new
'lib' file from our command line script so that we can test it from
the command line and make sure we didn't break anything. s3 upload
mechanics can now be used if we require that 'lib' file from anywhere
in our API and just needs to be passed a file.