Add crypto's randomBytes to decide filename
Include crypto in our process to create a random name generated by
crypto.randomBytes. This will prevent name-collision on AWS, so
there isn't a chance users using the same "name" overwrite other users'
photos with that name.
Since crypto does not return a promise, but IS async, we have wrapped
it in a promise which is resolved or rejected based on crypto's success
or failure. Did the same thing for s3.upload.
Tested this functionality by adding a .then/.catch block to
bin/aws-upload.js