Finishes API create function
Starts promise chain with s3Upload, then saves the necessary pieces
in mongo, then returns a response using that information. Refactored
lib/aws file to expect the object that multer creates (using
originalname, and the mimetype provided). Refactored the bin/aws
node script as well because we changed the contract of lib/aws.